Magic cards finished

+++ api/crud/organ.py
@@ -46,18 +46,16 @@ def player_has_organ(db: Session, player_id, tipo: str):
     else:
         return False
 
-def player_has_organ_to_cure(db: Session, player_id, tipo: str):
-    # Buscar el registro de organ
-    if(tipo != OrganType.magic):
-        db_organ = db.query(Organ).filter(Organ.player_id == player_id, Organ.tipo == tipo).first()
-    else:
-        db_organ = db.query(Organ).filter(Organ.player_id == player_id).first()
+def player_has_organ_to_cure_infect(db: Session, player_id, tipo: str):
+
+    db_organ = db.query(Organ).filter(Organ.player_id == player_id).first()
     
     if db_organ:
         return True
     else:
         return False
 
+    
 def player_can_steal(db: Session, player_id: int, player_to: int, tipo: OrganType):
 
     # Buscar el registro de organ
@@ -65,14 +63,15 @@ def player_can_steal(db: Session, player_id: int, player_to: int, tipo: OrganTyp
 
     db_organ2 = db.query(Organ).filter(Organ.player_id == player_to, Organ.tipo == tipo).first()
 
+
     if db_organ2:
         # Si está inmunizado, no puedo robarlo
-        if db_organ2.cure == 2:
+        if db_organ2.cure == 3:
             return False
         if db_organ:
             return False
         else:
-            return False
+            return True
     else:
         return False
 
@@ -95,24 +94,52 @@ def add_virus_to_organ(db: Session, player_to: int, card_tipo: str, organ_to_inf
             db.refresh(db_organ)
         # # # Si el órgano tiene una cura de su tipo (1), se le quita la cura si el tipo de órgano es igual al tipo de carta
         elif db_organ.cure == 1:
+            print("db_organ.tipo: ", db_organ.tipo)
+            print("card_tipo: ", card_tipo)
             if db_organ.tipo == card_tipo:
                 db_organ.cure = 0
                 db.commit()
                 db.refresh(db_organ)
+            else:
+                print("Error al añadir virus al órgano")
+                return False
         # # # Si el órgano no tiene cura ni virus, se le añade el virus:
         elif db_organ.cure == 0:
             if db_organ.virus == 0:
-                # Si el tipo de carta es magic, se le añade el virus (= 2)
-                # Si el tipo de carta es del tipo del órgano, se le añade el virus (= 1)
-                if card_tipo == OrganType.magic:
-                    db_organ.virus = 2
-            
-                elif db_organ.tipo == card_tipo:
-                    db_organ.virus = 1
+                if db_organ.tipo != OrganType.magic:
+                    if (db_organ.tipo != card_tipo and card_tipo != OrganType.magic and db_organ.tipo != OrganType.magic):
+                        print ("Error al añadir viruss al órgano")
+                        return False
+                    # Si el tipo de carta es magic, se le añade el virus (= 2)
+                    if card_tipo == OrganType.magic:
+                        db_organ.virus = 2
+                    # Si el tipo de carta es del tipo del órgano, se le añade el virus (= 1)
+                    elif db_organ.tipo == card_tipo:
+                        db_organ.virus = 1
+                
+                elif db_organ.tipo == OrganType.magic:
+                    # Si el tipo de carta es magic, se le añade el virus (= 2)
+                    if card_tipo == OrganType.magic:
+                        db_organ.virus = 2
+                    # Si el tipo de carta es del tipo del órgano, se le añade el virus (= 1)
+                    else:
+                        db_organ.virus = 1
+                        if card_tipo == OrganType.heart:
+                            db_organ.magic_organ = 1
+                        if card_tipo == OrganType.brain:
+                            db_organ.magic_organ = 2
+                        if card_tipo == OrganType.intestine:
+                            db_organ.magic_organ = 3
+                        if card_tipo == OrganType.lungs:
+                            db_organ.magic_organ = 4
+
                 db.commit()
                 db.refresh(db_organ)
             # # # Si el órgano ya tiene el virus, se elimina el órgano
             elif (db_organ.virus == 1) or (db_organ.virus == 2):
+                if (db_organ.tipo != card_tipo and card_tipo != OrganType.magic and db_organ.tipo != OrganType.magic):
+                    print ("Error al añadir virus al órgano")
+                    return False
                 #Eliminamos el órgano
                 db.delete(db_organ)
             else:
@@ -128,7 +155,7 @@ def add_virus_to_organ(db: Session, player_to: int, card_tipo: str, organ_to_inf
     return True
     
 
-def add_cure_to_organ(db: Session, player_id: int, tipo: str, organ_to_cure: str):
+def add_cure_to_organ(db: Session, player_id: int, card_tipo: str, organ_to_cure: str):
     # Buscar el registro de organ
     db_organ = db.query(Organ).filter(Organ.player_id == player_id, Organ.tipo == organ_to_cure).first()
     # Si existe el órgano:
@@ -138,39 +165,79 @@ def add_cure_to_organ(db: Session, player_id: int, tipo: str, organ_to_cure: str
             db_organ.virus = 0
         # Si el órgano tiene virus normal (1), se le quita el virus si el tipo de carta es igual al tipo de órgano
         elif db_organ.virus == 1:
-            if db_organ.tipo == tipo:
+            print("db_organ.tipo: ", db_organ.tipo)
+            print("card_tipo: ", card_tipo)
+            if (db_organ.tipo == card_tipo) or (db_organ.tipo == OrganType.magic) or (card_tipo == OrganType.magic):
                 db_organ.virus = 0
-        # Si el órgano no tiene virus ni cura, se le añade la cura:
-        elif db_organ.virus == 0:
-            # Si el tipo de carta es magic, se le añade la cura (= 2)
-            if tipo == OrganType.magic:
-                # # Si el órgano no tiene cura, se le añade la cura
-                if db_organ.cure == 0:
-                    db_organ.cure = 2
-                # # Si el órgano tiene cura, se inmuniza el órgano
-                elif (db_organ.cure == 1) or (db_organ.cure == 2):
-                    db_organ.cure = 3
-                else:
-                    print("Error al añadir cura al órgano")
-                    return False
-                
-            # Si el tipo de carta es del tipo del órgano, se le añade la cura (= 1)
-            elif db_organ.tipo == tipo:
-                # # Si el órgano no tiene cura, se le añade la cura
-                if db_organ.cure == 0:
-                    db_organ.cure = 1
-                # # Si el órgano tiene cura, se inmuniza el órgano
-                elif (db_organ.cure == 1) or (db_organ.cure == 2):
-                    db_organ.cure = 3
-                elif db_organ.cure == 3:
-                    print("El órgano ya está inmunizado")
-                    return False
-                else:
-                    print("Error al añadir cura al órgano")
-                    return False
             else:
                 print("Error al añadir cura al órgano")
                 return False
+        # Si el órgano no tiene virus ni cura, se le añade la cura:
+        elif db_organ.virus == 0:
+            if db_organ.tipo != OrganType.magic:
+                # Si el tipo de carta es magic, se le añade la cura (= 2)
+                if card_tipo == OrganType.magic:
+                    # # Si el órgano no tiene cura, se le añade la cura
+                    if db_organ.cure == 0:
+                        db_organ.cure = 2
+                    # # Si el órgano tiene cura, se inmuniza el órgano
+                    elif (db_organ.cure == 1) or (db_organ.cure == 2):
+                        db_organ.cure = 3
+                    else:
+                        print("Error al añadir cura al órgano")
+                        return False
+                    
+                # Si el tipo de carta es del tipo del órgano, se le añade la cura (= 1)
+                elif db_organ.tipo == card_tipo:
+                    # # Si el órgano no tiene cura, se le añade la cura
+                    if db_organ.cure == 0:
+                        db_organ.cure = 1
+                    # # Si el órgano tiene cura, se inmuniza el órgano
+                    elif (db_organ.cure == 1) or (db_organ.cure == 2):
+                        db_organ.cure = 3
+                    elif db_organ.cure == 3:
+                        print("El órgano ya está inmunizado")
+                        return False
+                    else:
+                        print("Error al añadir cura al órgano")
+                        return False
+                else:
+                    print("Error al añadir cura al órgano")
+                    return False
+            elif db_organ.tipo == OrganType.magic:
+                # Si el tipo de carta es magic, se le añade la cura (= 2)
+                if card_tipo == OrganType.magic:
+                    # # Si el órgano no tiene cura, se le añade la cura
+                    if db_organ.cure == 0:
+                        db_organ.cure = 2
+                    # # Si el órgano tiene cura, se inmuniza el órgano
+                    elif (db_organ.cure == 1) or (db_organ.cure == 2):
+                        db_organ.cure = 3
+                    else:
+                        print("Error al añadir cura al órgano")
+                        return False
+                    
+                else:
+                    # # Si el órgano no tiene cura, se le añade la cura
+                    if db_organ.cure == 0:
+                        db_organ.cure = 1
+                        if card_tipo == OrganType.heart:
+                            db_organ.magic_organ = 1
+                        if card_tipo == OrganType.brain:
+                            db_organ.magic_organ = 2
+                        if card_tipo == OrganType.intestine:
+                            db_organ.magic_organ = 3
+                        if card_tipo == OrganType.lungs:
+                            db_organ.magic_organ = 4
+                    # # Si el órgano tiene cura, se inmuniza el órgano
+                    elif (db_organ.cure == 1) or (db_organ.cure == 2):
+                        db_organ.cure = 3
+                    elif db_organ.cure == 3:
+                        print("El órgano ya está inmunizado")
+                        return False
+                    else:
+                        print("Error al añadir cura al órgano")
+                        return False
         else:
             print("Error al añadir cura al órgano")
             return False
